Semi-wild stab at bug 34601: 'drop media file to donate here' sometimes silently fails in UploadWizard
The hidden file input box gets positioned over a button that it's covering, since you can't just style the input to something sane.
For some reason, sometimes this initial position is wrong and totally off from where it belongs.

This might be due to some weird initialization order thing; in theory all the CSS should already be loaded and applied before we start running any of the JS in the same module so we don't expect it to be that. Could be some unrelated bit of behavior, though.

Attempting to reposition the file input shortly after we initially do, see if that helps any.

Change-Id: Ifd99649f6ec2603edb23afb80e8ee1450469c2fe