May or may not help with bug 34601 UW mispositioned file control - still can't reproduce this on a local install.
Force UploadWizard styles to load early.... there *is* HTML at output time, though it's "hidden" until some additional styling is applied to it.
Sometimes in Chrome some of the HTML flashes briefly without style, which seems to be when we pick up the wrong positioning... this hopefully will avoid that possibility by ensuring the styles are preloaded.