(bug 33337) Large Thumbnail Preview popup has scrollbar
Dialog box generated has widht and height set to 'largeThumbnailWidth' and 'largeThumbnailMaxHeight' variables set in UploadWizard.config.php file.

As dialog titlebar took up some space, hence the total height to display the preview was getting reduced.
Fixed this by setting the dialog to have minHeight and minWidth according to the config variables, and then expand according to the preview generated.
Tested it in Chrome 18 and firefox 10.

Change-Id: Id99304cab3ef9aade83ab39c828d81a60e4e56de