(bug 32246) Weird behavior when clicking on the name or the thumbnail of the file
Click on the filename was intercepted by the hidden input box
and not the filename div. Fixed this by giving the filename container
div a higher z-index than the input container.
Now nothing happens when clicking on the filename div and a preview
popup dialog opens on clicking on the thumbnail of the file.

Change-Id: Iafbbf64db535a0a2a717d3e7dfb56a5bfe370925