(bug 35870) Fix pre-upload preview on IE 10 for Special:UploadWizard
When loading thumbnails, we fetch the actual file data so we can find EXIF data in JPEGs
and do suitable rotation fixes.

This currently uses binary strings via FileReader.readAsBinaryString, but that's not
implemented in IE 10.

Workaround uses FileReader.readAsArrayBuffer and converts to a binary string in case
FileReader.readAsBinaryString is missing but readAsArrayBuffer is there.

Change-Id: If9a89f4adf60db6b533566b58a5cb873b9dcb8cd