Experimental opt-in mode for UploadWizard's chunked uploads feature
We need to test this feature in production, but don't want it on
unconditionally in case there are more problems with it.

Setting $wgUploadWizardConfig['enableChunked'] = 'opt-in' will add
an "experimental" preference under the 'Uploads' tab, under the licensing
settings, which can enable the chunked mode per-user.

Setting it back to false or true later will remove the experimental
preference.

Change-Id: Iaca08db75cda2e7b25c7d7c032ead0b81c345e40