Generate client-side previews for GIF and PNG files.
Previously we checked for the presence of JPEG metadata, which prevented
GIF and PNG thumbnails from being generated. This is unnecessary as
the makePreview() function already checks the isPreviewableFile()
function, which only returns true for PNG, GIF and JPEG files.

Patchset 2: when adding multiple files, only add "Show thumbnail"
control if the file would be previewable client-side.

Change-Id: I1f81d8bd37c959a3f1b0279b559ef55a6a6cdc47