(bug 32843) Don't call the success method for checkFile() if there's an error
For whatever reason, there was no effort to stop the success method from
being called when a new file input was loaded. This patch adds a field to
the UploadWizardUpload object, accessible in the interface file via
this.upload.hasError, that indicates what kind(s) of error(s) are found
in the input. The method *does not return*, so it is possible to
override or ignore an error by setting this.upload.hasError[type] to
false inside of an error method.

If it makes more sense to just exit, that would be a quick fix.

Patchset 2: Non-FileApi browsers wouldn't get this patch until now,
            this patchset fixes it by checking in multipl places.

Change-Id: I22d9885f8153db9aa634ad4143f5b3430506750d