(bug 26167) Start uploading as soon as a file is selected
Rather than waiting for a form submit  to upload the file,
this patch will just start uploading as soon as the file
input is filled. This may or may not be beneficial behavior,
but at least it works.

Patchset 2: On a remove event, re-check the uploads and
            possibly show the next button.

Patchset 3: Prevent user from accidentally leaving the page
            while uploads exist.

Patchset 4: The long-awaited shift to using makeTransitioner.
            It properly uploads 3 at a time, now.

Patchset 5: Remove a superfluous call to showNext, and rebase
            on top of master.

Patchset 6: Add in a configuration parameter to disable the
            new feature, and replace the removed things to
            make the old way work again.

Patchset 7: Make a change to how checkFilename works, so PNGs
            can be properly uploaded with this (and any other)
            method.

Change-Id: Ia4214b1bb0a139f09d5ac3c50ec58e9041550109