Use user preference instead of cookie to permanently skip tutorial step.
Thanks to the new options API in MediaWiki 1.20, we can permanently dismiss
the introductory tutorial step by setting a user preference. As a general
UX principle, we should inform the user about how to change their
preference later; we do so here by means of a tipsy tooltip on the
checkbox.

Patchset 2: made prefs link go straight to upload tab

Change-Id: I835ebaff5e39892ad6d304fcd4d067a314519d80