(bug 32469) Remove alert message and use callbacks for validation
Until now, UW has had a set of methods called detailsValid, valid,
and getWikiText. Those were all called in order to get information
about the details in the form, and returned booleans and strings,
respectively. This, however, caused problems--if one or more of
the fields was, for whatever reason, not ready to be validated,
the validation methods returned false or null and gave the user an
alert message--not very modern, and a bad experience for the user.

This patch removes at least one alert message, which is a win, and
shifts the valid, detailsValid, and getWikiText methods to use a
callback system, which should make the interface slightly more
responsive, and make the validation system slightly more
JavaScript-standard.

Change-Id: Id262bbe4533f660054923f2705033d2df8684b80