(bug 36840) Reset the uploads count in the reset method
When circling the process a lot, it's helpful to have the "upload
more files" button, but whenever you click that and try to upload
multiple files, there is an error, because the count of uploads
does not get properly reset after the click event is fired.

This patch adds in a line that properly resets the count, and
fixes the problem.

Change-Id: Iccdb003d8ff141f9fbe31b4b7448135a09684b85