(bug 37039) Remove complaints based on was-deleted warnings
The API will send back the was-deleted warning if the filename
being submitted used to exist, but was deleted. Until now, UW
would complain about it, pretty much ending the upload process.

This patch will ignore that warning, if it's the only one.

Also, we get to remove an extra API call. Good day!

Patchset 9: Fixing imageinfo call

Change-Id: I119703f1f2b70725f491e750e856cc3f320d7e5d