(bug 37108) Add the current time to the filename in FormData
FormData sends the server *exactly* the filename given to it. This
is not helpful, because we don't care what the filename is on the
first step, since it might change later (and it had better, if we
get a duplicate error). But if the file's contents exists, we
can't fix that at a later step, so we'd better detect that error.
Turns out, the server will only return 'duplicate' if it finds
both, so we avoid that dilemma by giving the server a filename it
could never have seen before.

Change-Id: I1138f8f0394963d27552fc046b96200dad45ced1