(bug 30625) Ignore warnings, don't assume fatal if success
This is meant to be deployed with an API patch, but should work on
its own. It simply ignores warnings if the upload result was
"Success", unless they are important ones (duplicate, was-deleted,
exists, and the like)

Without the API patch, however, this will do nothing, since no
warnings will be returned in the relevant place.

Patchset 5: Slightly improved comment

Change-Id: I14d961af3ba7b9eb553df4aef9e868137354ce8a