(bug 37150) Add in a check for FileApi availability
The current way we check for FormData availability is all well and
good, but if we rely on only that, it causes problems. Checking for
FileApi functionality is also necessary....as Safari is not Firefox
less-than-four, but it does _not_ have FileReader support. This
patch checks for the proper funcitonality, and should fix the bugs
in Safari.

Change-Id: Ifc12bcb9f03a7a166f673298ccdabbecf695dfd4