Fix a problem with the start-immediately patches
In order to make the start-immediately functionality work well, we
had to add in a call to showNext in the remove method of the
Upload class. However, that method is called both when the user
removes an upload _and_ when it gets removed as part of the normal
cycling procedure. However, when every upload is removed, showNext
causes unrecoverable errors. This patch adds in a reasonable test
to stop this from happening, fixing a serious usability problem.

Change-Id: I7e08237c0670e049b8f448c4162edba0a4f12d9b