(bug 31189) Add languages that exist and have a template
This might not be desirable (could use some UX help), but it adds
in any supported languages that have a template associated with
them. For wikis that don't have all the language templates, this
will likely result in fewer supported languages, but for Commons,
this will bring in a lot more options. It's very likely that we'll
need to talk about the implications of that before deploying this
change, but it definitely works now.

OK, there were some tricks I didn't realize existed. And apparently
the Title class was even more complicated than I originally thought
it was. That's annoying.

Thanks for the review, Kaldari!

Patchset 11: Just whitespace fixes.

Change-Id: I3f34e18499b11c717c76c68ddbcdf492ec6cadf3