(bug 33330) Add 'deed.__' to the license URLs
This just adds in the translated version of any deed for the
language code currently being used by the user. This works because
creativecommons.org will sanely fall back to the languages it
knows if we link to a language it doesn't--so anyone should be
able to full use the CC links now without trouble.

Change-Id: I4005b0573ae7e6e86973f330a21470d6b9a416ce