(bug 38022) Check for completeness in confirmCloseWindow
If you want to close the window, we stop you. We want to protect
your data from being lost by an accidental "Ctrl-W" when you meant
"Ctrl-V" or whatever. But on the last page ("thanks")...there's no
longer any danger. Your data is saved. So let's allow total
anarchy.

This patch fixes the bug and generally makes for a better system.
It should also be modified to work with change I4994d2bf because
it uses a .length === 0 check. We need to account for empty
uploads, which this patch does only partly.

Change-Id: If287818dded49aebfe6b9724856f729a26905c6a