(bug 37073) Make sure copyMetadataCrtlDiv exists and is correct
Because it is possible to remove uploads after the details interface
has been created, sometimes copyMetadataCrtlDiv was not present
anymore. This delays the creation of the interface until it is needed.

This fixes a JS exception that caused the Next button not to show up
for users, a bug reported once a day on UW feedback page.

PS2: Adding bug ticket to commit msg.
PS3: Add a function description and a ;

Change-Id: I28fb937e1af1faa072c4452655fe2964af2c21ed