(bug 35520) Don't wikilink input that is already likely to be wikicode
The author field is normaly used as text for the wikilink to your
username. Many users, however, have their own ways to form their
attribution.

This will use raw wikicode for the author field if the user adds
'[' or '{' in the author field. If they don't, the field input
will still be used as the text for the wikilink to the username.

Change-Id: I0a6580e953754b290abd0844b2bb95e2f62c0333