Respect languageTemplateFixups
If languageTemplateFixups are defined, they need to be respected when
determining the langauges for which there are language templates.

When I installed UW on my private test wiki, which doesn't have tgl, but
which does have tl (from Commons, "templatelink" template), Tagalog
ended up in the language list, even though there's no template for it.

Change-Id: I396aa758608b6bbc66162ab2ef2ce356ebfecf41