Correct the error count
the error count is off otherwise. Comment on that function says it was a
bit hacky. This fix doesn't improve on that, but at least it gives me
sensible error counts.

To see the problem: choose two files to upload, select "provide
individual licenses", the just click next on the describe step without
entering descriptions or choosing a license. I got a message saying
"There are 2 errors..." when the interface clearly showed me four.

With this fix, the error count did in my manual tests match the actual
number of errors displayed on the page.

Change-Id: If0cb1ff2da002b62e1020d68d37d4c5fdd2e08e4