(bug 43593) error for restricted flickr content.
Restricted content from flickr needs authorization. Since currently there
is no way to find out if the content is restricted or invalid we try to
display a single error message for both conditions.

Todo: Find out a way to know if the flickr set is restricted and authorize users to fetch its content.

Change-Id: Id906ae3f205ed4742e2e03cc1faa49171e5d47c9