Mark required fields under "Release Rights"
Mark the required fields under "This file is not my own work" in
the "Release Rights" stage as required in the same way as the fields
in the "Describe" stage.  (By adding a key "* = This field is
required." and by adding red asterisks to the field labels.)

Bugzilla Bug 42775 - Confusing annotation about mandatory fields
in Special:UploadWizard

https://bugzilla.wikimedia.org/show_bug.cgi?id=42775

Note that the radio choice between "This file is my own work"
and "This file is not my own work" is also in some sense a
required field but as it has no single label as the text box
fields do I haven't marked it as required in any way.

Under "This file is my own work" there is another field (the
input for username) that has the same problem - it is required
but there is no obvious label to put an asterisk by.

I have left these problems unsolved and they would make sense
as future work.

One possible problem with this actual change is that the key appears
in the "Release Rights" section initially when all required input
fields are hidden and thus no labels with asterisks are visible.
Might this be confusing to the user?

Change-Id: I6c0274f9d10bd83d5916a475458b0cc88ddc81c2