Replacing deprecated $j.browser with $.client
jQuery.browser is deprecated and already removed in the latest versions
of jQuery. Since feature detection does not work here, we still have to
rely on browser-sniffing. For this purpose, we use $.client, which is
already used in mw.UploadWizardInterface, thus no additional dependency
is required.

This also eliminates a bug in this code where the version is compared
against a string which leads to confusing and undesired results (cf. Bug
43821)

Bug: 43821
Change-Id: I223dfd7857bae9ae904dcd5b9232b64338f8cd4c