Improved file-control positioning
Currently the UploadWizard sets 2 Intervals per file that are fired
every 500ms accessing the DOM. You can imagine that this can lead to
high CPU load at the client. But these intervals are even not cleared
after they are not required anymore and make it impossible to upload 50
files at once on old, slow machines. (In a test on a Intel Pentium IV,
1.7GHz, Firefox 11, Windows XP SP3, I was unable to upload more than 10
files without running into an unresponsive UI.)

This commit eliminates stupid "polling" and, instead listens to
resize-events. It also avoids caring about uploads that were removed by
the user or about file inputs that are filled and are just kept for
accessing the file contents. All this leads to less CPU load, a more
responsive UI and finally to less pollution of the environment and more
happy users.

A call to $.client.profile() was removed because the variable that
receives the return-value seems to be unused.

Bug: 37997
Change-Id: Icd8642c663869e56c2397b69931f9ade4bf5bf7c