Use checkboxShiftClick for "Copy Metadata" checkboxes
With CheckboxShiftClick functionality, the user can change the state of
multiple checkboxes at once by pressing the shift key. This
functionality is enabled by default for all checkboxes that are present
when mw.hook('wikipage.content') is fired but not for checkboxes like
these that are dynamically generated using JavaScript.

Another way to achieve this would be firing the hook as soon as Upload
Wizard makes changes to the HTML DOM:
   mw.hooks('wikipage.content').fire( <content that was newly created> );
This would have the downside that other scripts could interfere with
UploadWizards controls in unforeseeable ways.

Also did some cleanup:

* Marking instances of jQuery with $-prefix
* Declaring variables in the beginning of the function.

Change-Id: I97f95e10bcfd03105a7907d95c10a78d944083db