Update and remove unnecessary messages
Remove default "Thanks for using our new upload tool!"
message as well as the translation link, as the tool is
neither new or in dire need of translation. A custom
welcome message can now be added with headerMessage in
the config if one desires.

Also, update i18n for mwe-upwiz-thanks-explain to be
more general.

Bug: 36149
Change-Id: I5fa30bf6b9690511e1dcaaa66d944f412672aa7b