UploadWizard: Add support for PostgreSQL
The extension-specific database schema updater had MySQL-specific syntax
for the uw_campaigns table, so create an alternative update script for
PostgreSQL.

This allows the update.php script to complete successfully under
PostgreSQL when the UploadWizard extension is enabled.

The uw_campaigns table is only used in advanced cases.  The
correct behavior of the resulting table under PostgreSQL has not been
tested.

Bug: 64067
Change-Id: I48e91b9f17e7a3142a5ffe718b56848d1f90b7df