UploadWizard: Check for duplicate titles
Upload Wizard allows uploading multiple files within one "batch".
However, it is easily possible that a user specifies the same name
for two or more files to be uploaded. Until now, Upload Wizard
attempts to upload all files. Most of the time, the API and
database get it right and return an error. The user can then
specify a new title. But maybe, the user would have preferred
giving a title of one of the already uploaded files to one of the
remaining files for which an error was returned. Therefore the
correct behaviour is to make the user aware of that issue before
uploading any of the files.

If the database and API do not get it right; for example due to
replication lag or other weird issues (consider that the files are
submitted either simultaneously or in a short interval), Upload
Wizard will overwrite files like experienced in bug 64883.

Further changes:
- Add classes to error labes indicating when validation happens.

Bug: 64883
Bug: 54750
Change-Id: Ic57d7c9bb4f986f3298a57f9f0285567a076777e