Feature is not reliably browser-testable; remove
The preference-setting happens as a side effect, and there's no
way to tell Selenium that it succeeded or failed. So, sometimes,
when we check the preferences to see if it's been set, we're
too early and it hasn't.

We don't want to rewrite the feature so that it blocks the user
from getting to the add files step.

So we'll have to remove the test.  :/

We might be able to test this better with a mock API in unit tests.

Change-Id: I4e4a423b4da89d4c00780457236fff1c446960c7