Remove reservedIndex
Once upon a time, I embarked on a multiple-day refactoring journey to accomplish
this. Now, I apparently can just delete all the instances of it. I tested this
and apart from reordering the images a bit, there doesn't appear to be any ill
effect, and you know what? I'm OK with people being *slightly* confused by the
ordering if it means we can work with this codebase without fearing for our
lives every time.

Change-Id: Ifc55aac4b8b14e24e59c7a015282cc3dc4e25c6d