Add "almost finished" message and remove parseInts
The "Finished!" message was appearing too early, so let's make
it totally clear here. Message will tell the user that the server
is processing the images or something...we aren't always sure that's
what's happening, but it's a believable and comforting fiction.

And of course, "parseInt" is a stupid way to make things into integers,
and a stupider way to round things down. So we'll stop that.

Bug: T62483
Change-Id: If41718dd8d0399f16ab06627357a0ce238e4f986