[UploadWizard] Introducing the usage of an object reference
The parameter objref will be used to have a bot updating the
lists of monuments with a thumbnail image of one of the uploaded
images.

This patch also adds the keyword 'useObjref' to the home button
and the keyword 'dropObjref' to the begin button configuration.

If it is set and there is an objref parameter passed with the request,
clicking the home button leads to the page passed as the reference (this
may be any page in any wiki defined for interwiki links; it is intended
to be the referer, though). If the begin button is configured to drop the
object reference, it will reload the Upload Wizard page, but will drop
the listUpdate parameter from the URL.

Independently from the button configuration, if the objref parameter is
set, the Upload Wizard adds the reference parameter to the description
page inside a comment tag. This will later be used by a bot that
supplements an image thumbnail to monument lists.

Bug: T105281
Change-Id: Ia974a148a6326dbca17a70c54a0cffcd2e4a39ac