[UploadWizard] Introducing the usage of an object reference
This patch adds checkboxes to the uploads on Upload Wizard's upload page.

Users can then decide, to which image the object reference should be
applied. Whether checkboxes are being displayed depends on the objref
parameter passed to the Upload Wizard. The checkboxes are only enabled
if the parameter updateList contains any value. If the parameter is not
set or not given the checkboxes are disabled and a notice is being
displayed stating that the list entry already contains an image.

Bug: T105283
Change-Id: I6012a7f66267cc361e34ccd30bfcb06961701a83