Do not claim that a file with given name already exists if the extension differs
MediaWiki's action=upload API will return the 'exists' warning only if
a file with exactly same name exists, and the same 'exists-normalized'
for all kinds of "similar" names. If you're uploading 'Foo.jpg', it
will return 'exists-normalized' both if a 'Foo.ogg' exists (which
should not block upload), and if 'Foo.JPEG' exists (which should
block).

This is especially annoying when uploading an SVG version of a raster
drawing.

Double-check the API result and only block the upload if the extension
does not differ in a non-trivial way. For different extensions let the
upload happen.

Depends on I1cf240b991dc170c46ba76b207b4e4ba77f87956 in MediaWiki core.

Bug: T48741
Change-Id: I664a801dc20b1dab916a1669455bc5913121b92f