Do not obfuscate class names when adding prototype methods
This common pattern:

    var FP;
    function Foo() {}
    FP = Foo.prototype;
    FP.bar = function () {};

Guarantees that a programmer debugging the code and looking at the
call stack will have no idea what class the 'bar' method belongs to.
Works especially well when the short prefix is common to a number of
classes that all call each other.

    function Foo() {}
    Foo.prototype.bar = function () {};
    global.ns.Foo = Foo;

When your coworkers are suspicious of the previous one, this pattern
achieves almost the same effect, while being explainable by
"performance reasons". Works especially well when classes by the same
name appear in a number of namespaces, e.g. when using something akin
to MVC structure.

Change-Id: I37c22eb043e0f288a26c69898ace80478cf01d58