Split off uw.DateDetailsWidget from UploadWizardDetails (introduce uw.DetailsWidget)
This change adds uw.DetailsWidget and uw.FieldLayout.
* uw.DetailsWidget represents a single logical field in the "Details" step.
  It can be composed of multiple smaller widgets (e.g. a "location" field
  could be composed of "latitude" and "longitude" inputs).
* uw.FieldLayout is a small extension of OO.ui.FieldLayout to magically
  display error and warnings messages about user input in uw.DetailsWidget.

With these we should be able to move away from jquery.validate, which has
been causing no end of trouble.
* uw.ui.Details has a small hack added to know about new-style error
  messages.
* uw.controller.Details refactored to validate both old-style and
  new-style fields before submitting the form.

As a first step of migration towards the new system, I converted the date
field, which was already using OOjs UI. The only user-visible change is
that the error message about the field being required looks different.
* uw.DateDetailsWidget is a new class implementing the UI (code moved from
  UploadWizardDetails), the validation rules, the copying rules and
  getting wikitext from user input.
* UploadWizardDetails only has to create the field, put it in the form,
  and get the value back. (And prefill the initial one from EXIF.)

This change does not touch the hint system, using jquery.tipsy. While
OOjs UI has the code for this in place, I think it would be confusing
to users to have two very different-looking tooltip systems side-by-side.
We'll change this later.

Bug: T96917
Change-Id: I446b5b3e8ddb07fc31bbad77ca57ba50b7f96082