Split off uw.CategoriesDetailsWidget from UploadWizardDetails (improve warnings)
This change converts the category fields to use uw.FieldLayout and
uw.DetailsWidget rather than custom code. As this field was already using
OOjs UI, there should be no user-visible changes.

It also generalizes the warning about submitting uploads with no
categories to use the warning system from uw.DetailsWidget.
* uw.controller.Details now knows about the warnings and displays the
  confirmation dialogs if there are any, rather than hardcoding the
  category check.
* Some code has been moved to UploadWizardDetails so that the controller
  doesn't have to know about individual fields anymore.

The warning message in the form is temporarily hidden with CSS, though,
to avoid changing the behavior in only some parts of the form until we
change other things to use similar warnings. I filed T117117 and T117119
to track this work.

Bug: T96917
Change-Id: Iee797ce14e2457759e40e08dcf16746e7041384c