Remove dead code dealing with 'unknown' language
This looks like it was meant to handle the case where the wiki had no
language templates defined, and thus config.uwLanguages is empty.
However, other code already badly fails with exceptions in that case.
If we want to support this, we should just not show a dropdown at all.

Change-Id: Ida3567c38717e30afc15a0eebb1daabca863d8ae