Split off uw.DescriptionsDetailsWidget from UploadWizardDetails (and OOUI-ize)
This change converts the descriptions field to use OOjs UI, moving the
code to uw.DescriptionDetailsWidget and uw.DescriptionsDetailsWidget.
* uw.DescriptionDetailsWidget implements logic and UI for a single
  language and description pair. This code was previously scattered in
  mw.UploadWizardDescription, uw.model.Description and mw.LanguageUpWiz,
  all of which are removed now.
* uw.DescriptionsDetailsWidget implements logic for managing multiple
  descriptions and adding/removing them, which was mostly in
  UploadWizardDetails.

The dropdowns and text fields now use OOUI widgets and are aligned on one
line (as they used to be a long time ago, but this broke at some point).

Everything related to uw.model is also removed, since that attempt at
refactoring seems to have been aborted.

Some weird things that I noticed, but did not change:
* There is some questionably-looking logic changing some languages to
  others (now in uw.DescriptionDetailsWidget). I'm not sure how much
  of it is needed.
* 'descriptionlang' and 'description' URI query parameters are handled
  differently.

Bug: T96917
Change-Id: I9563ff3bf7acb24ab0df35930643df81aee89dba