Revert "Define required 'defaults' field in licensing config"
Gergő pointed out in code review on the original patch how it's really
silly that this field is an array when only the first value is used.
Turns out that CampaignSchema actually specifies it as a string,
and will prevent Campaign pages from being saved when it's an array...

On the bright side, that means it was actually impossible to use in
campaigns (when defined as an array, validation would fail and not
save the page; when defined as a string, UploadWizard would throw
exceptions), so we'll be able to correct the code to use it as a string
without breaking changes.

This reverts commit 0c142e5d6c520d45877a822da6d67499396db428.

Bug: T123840
Change-Id: I5006577a163e181d3b73fc4fc8895443ec0cf6d7