Avoid parsing HTML when creating <input> nodes
We used to create them like $( '<input type="checkbox" />' ), which
actually does HTML parsing under the hood, because on IE 8 and lower
trying to change the 'type' of an <input> node throws an exception.
But we dropped support for IE 8 recently and no longer need to do that.

Didn't touch any places where we otherwise do a lot of HTML parsing.

Change-Id: Ie136c427c49d83bb6aefe49b9aaad9dbec65ef16