mw.UploadWizardUpload: Disentangle JPEG metadata from filling uploads
As far as I can tell (it's hard to tell because it's entangled), we
only use this to render the thumbnails with the right rotation and
aspect ratio. So let's just do it as part of the thumbnail generation.

This should let us make other stuff less horrible. In particular, this
means that the this.uploads array now holds uploads in the same order
in which they were created (and in which they're shown in the interface),
which in turn results in UW uploading files in the same order, and
which allows us to remove some crappy sorting from elsewhere (added in
a37bba8dbc2e87b108c7e22c7ef0dabe847f07c5).

Change-Id: Ie9c0098074320a1fb9e63acab5672965f37a4700