uw.controller.Step: Emit 'load' earlier, so that mw.UploadWizard#removeErrorUploads works
We have this event handler defined in mw.UploadWizard:

	deeds: new uw.controller.Deed( this.api, this.config )
		.on( 'load', function () {
			wizard.removeErrorUploads();
		} ),

But here, before we we emitted 'load', we did this:

	$.each( this.uploads, function ( i, upload ) {
		if ( upload !== undefined ) {
			upload.state = step.stepName;
		}
	} );

This would clear the 'error' state from each upload (replacing it
with 'deeds'), and so none of the errored upload would be removed.
They would then inevitably fail in the Details step, where they
can't be removed.

This is very brittle, maybe we can do this in a different way later.

Change-Id: I89095f07e97fed9255a43ecf3a9191b541b51dab