Use local this.config rather than mw.UploadWizard.config where we need to
None of these classes have a dependency on 'ext.uploadWizard' (it
would result in circular dependencies), and so they are not allowed to
access mw.UploadWizard. We already pass the config to them when construcing.
This was causing unit tests to fail when not ran in specific order.

Also remove checks for `mw.UploadWizard.config` itself, which were
presumably added when tests started failing...

Change-Id: Ib38374ae8f444c05b46362d14f5be1620f8aee14