Remove the ability to configure enableFormData, enableMultiFileSelect, enableMultipleFiles
All of these are core features of UploadWizard. Perhaps they weren't
years ago when the project was started, and when browsers did not
always support the HTML and JS features necessary to implement these,
but they are now. I don't think anyone has tested when happens if any
of these is disabled for years.

Bug: T126763
Change-Id: Icb0a59e4ada76f253041ca224b23b8c3796eb41d