Rework TitleBlacklist integration
Remove 'useTitleBlacklistApi' from config, which used a fragile check
depending on the order in which TitleBlacklist and UploadWizard
extensions were loaded.

Instead, to tell if TitleBlacklist is available, just try loading the
'mediawiki.api.titleblacklist' module. If it fails, it means we don't
have TitleBlacklist.

Remove the conditional dependency on that module, since it's loaded
through mw.loader.using() now.

Change-Id: I0a859b44c41925ad2d8b05f1d186b563498261e1