mw.DestinationChecker: Check for non-exact extension matches too
This is uglier than I hoped it would be. We have to fire two API requests,
since there's no way to check both for duplicates in a foreign repo and
for duplicates with a different extension. (And there's no way at all to
check for duplicates with a different extension in a foreign repo, but we
don't really care about these, we just want to stop the user from
accidentally shadowing the foreign file.)

Since the matched title can now be different from the inputted title,
changed uw.TitleDetailsWidget to use the matched one, if any.

Bug: T132356
Change-Id: I0a29d9ff20fe71fd9557feb1430164afdff9a151