Set 'group' => 'ext.uploadWizard' for all our modules
This way all of our code will be loaded in a single load.php request,
and no other code will be loaded in the same request. This should
allow us to make some sense out of the line and column numbers for
logged JS errors, and maybe resolve stumpers like T139130 or T136237.

Previously, we only set that for the 'ext.uploadWizard' module itself,
which was nonsensical. All of the dependencies would load together
with other code, then the final module would load separately.

Bug: T139130
Bug: T136237
Change-Id: I2d5ee1e4219c260af7d919c9843758e07b600df3