mw.FormDataTransport: Work around call stack limits for chunked uploads
In jQuery 2.x, nested promises result in nested call stacks when
resolving/rejecting/notifying the last promise in the chain and
listening on the first one, and browsers have call stack limits low
enough that we previously ran into them for files around a couple
hundred megabytes (the worst is Firefox 47 with a limit of 1024
calls).

Write some custom logic without chaining (just manual calls to
done/fail/progress and resolve/reject/notify), so that we never have
to go through hundreds of promises and thousands of function calls
when finishing the upload and resolving a promise.

Bug: T130610
Change-Id: Ia34134f4aac58d80a2752efe8103fdf3646e6078