mw.FormDataTransport: Don't set filekey with offset=0 when retrying
I am confused by this. I think this could happen if the API
returns the 'filekey', but doesn't return a 'result', causing
mw.FormDataTransport#uploadChunk to retry the request. But
I don't see in what case it would do that for uploads to stash.
We'll see if this stops the errors.

Bug: T143134
Change-Id: Ida72a5040176adbb6456582fc82596095ed0e835