Display warnings in confirmation window
These can be confusing. Categories, for example, is a
non-required field, but will yield a warning when trying
to submit empty. Up until the point the form is submitted,
a warning that it should be filled out has never been
displayed, and that confirmation window will be confusing.

All this does is also show a list of the warnings in the
confirmation window, which hopefully helps with the confusion.

Bug: T146155
Change-Id: Ibe7d22b68da0a1db0aa8672229e1f127de359907