mw.UploadWizardLicenseInput: Parse license templates as if they were used on a file page
Some of Commons' license templates, for example:
https://commons.wikimedia.org/wiki/Template:Images_by_Ulrich_Kalkschmidt
only appear when used on a page in the 'File' namespace.
Provide a 'title' with the action=parse query to avoid false rejections.
We don't know the real title, so just use a fake one, it shouldn't matter.

Follow-up to 5861fa0f27387bbb29d9db5c7e19996618d0691e.

Bug: T140903
Change-Id: Iad3647ace8ed7acac6a25488de94ddde51e1bd40